Update advanced.js Save and Apply
Issue: When clicking "Save and Apply" in the UI, the service was not restarted on some devices. This was probably because we were using the method callInitAction, which did not work reliably on all devices.

Solution: The code was adjusted to use fs.exec_direct to directly execute the init.d scripts (/etc/init.d/qosmate restart). This method works reliably on all devices and ensures that the service is correctly restarted when "Save and Apply" is clicked in the UI.